refactor(api): split activate from user service
